changed all the i64 to i32

+++ src/composite_fns.rs
@@ -5,13 +5,13 @@ use crate::replacing_fns::*;
 /// Check Args: interval -------
 /// Replace Args: interval_time, difference (interval)
 pub fn gliss_on_interval(
-    check_args: Vec<i64>,
-    replace_args: Vec<i64>,
+    check_args: Vec<i32>,
+    replace_args: Vec<i32>,
 ) -> (
-    (Box<dyn Fn(Vec<i64>, i8, u64, i8, u64) -> bool>, Vec<i64>),
+    (Box<dyn Fn(Vec<i32>, i8, u32, i8, u32) -> bool>, Vec<i32>),
     (
-        Box<dyn Fn(Vec<i64>, i8, u64, i8, u64) -> (Vec<i8>, Vec<u64>)>,
-        Vec<i64>,
+        Box<dyn Fn(Vec<i32>, i8, u32, i8, u32) -> (Vec<i8>, Vec<u32>)>,
+        Vec<i32>,
     ),
 ) {
     (
@@ -24,13 +24,13 @@ pub fn gliss_on_interval(
 /// Check Args: interval, lower_limit, upper_limit, interval_time
 /// Replace Args: interval_time, initial_direction (interval)
 pub fn group_on_interval_in_range(
-    check_args: Vec<i64>,
-    replace_args: Vec<i64>,
+    check_args: Vec<i32>,
+    replace_args: Vec<i32>,
 ) -> (
-    (Box<dyn Fn(Vec<i64>, i8, u64, i8, u64) -> bool>, Vec<i64>),
+    (Box<dyn Fn(Vec<i32>, i8, u32, i8, u32) -> bool>, Vec<i32>),
     (
-        Box<dyn Fn(Vec<i64>, i8, u64, i8, u64) -> (Vec<i8>, Vec<u64>)>,
-        Vec<i64>,
+        Box<dyn Fn(Vec<i32>, i8, u32, i8, u32) -> (Vec<i8>, Vec<u32>)>,
+        Vec<i32>,
     ),
 ) {
     (
@@ -43,13 +43,13 @@ pub fn group_on_interval_in_range(
 /// Check Args: interval, interval_time -------
 /// Replace Args: interval_time
 pub fn trill_on_interval(
-    check_args: Vec<i64>,
-    replace_args: Vec<i64>,
+    check_args: Vec<i32>,
+    replace_args: Vec<i32>,
 ) -> (
-    (Box<dyn Fn(Vec<i64>, i8, u64, i8, u64) -> bool>, Vec<i64>),
+    (Box<dyn Fn(Vec<i32>, i8, u32, i8, u32) -> bool>, Vec<i32>),
     (
-        Box<dyn Fn(Vec<i64>, i8, u64, i8, u64) -> (Vec<i8>, Vec<u64>)>,
-        Vec<i64>,
+        Box<dyn Fn(Vec<i32>, i8, u32, i8, u32) -> (Vec<i8>, Vec<u32>)>,
+        Vec<i32>,
     ),
 ) {
     (
@@ -62,13 +62,13 @@ pub fn trill_on_interval(
 /// Check Args: interval, interval_time -------
 /// Replace Args: glissnote_interval (from the 1st note), gracenote_duration
 pub fn gracenote_on_interval(
-    check_args: Vec<i64>,
-    replace_args: Vec<i64>,
+    check_args: Vec<i32>,
+    replace_args: Vec<i32>,
 ) -> (
-    (Box<dyn Fn(Vec<i64>, i8, u64, i8, u64) -> bool>, Vec<i64>),
+    (Box<dyn Fn(Vec<i32>, i8, u32, i8, u32) -> bool>, Vec<i32>),
     (
-        Box<dyn Fn(Vec<i64>, i8, u64, i8, u64) -> (Vec<i8>, Vec<u64>)>,
-        Vec<i64>,
+        Box<dyn Fn(Vec<i32>, i8, u32, i8, u32) -> (Vec<i8>, Vec<u32>)>,
+        Vec<i32>,
     ),
 ) {
     (
@@ -80,13 +80,13 @@ pub fn gracenote_on_interval(
 /// Check Args: interval, interval_time -------
 /// Replace Args: glissnote_duration
 pub fn approachgliss_on_interval(
-    check_args: Vec<i64>,
-    replace_args: Vec<i64>,
+    check_args: Vec<i32>,
+    replace_args: Vec<i32>,
 ) -> (
-    (Box<dyn Fn(Vec<i64>, i8, u64, i8, u64) -> bool>, Vec<i64>),
+    (Box<dyn Fn(Vec<i32>, i8, u32, i8, u32) -> bool>, Vec<i32>),
     (
-        Box<dyn Fn(Vec<i64>, i8, u64, i8, u64) -> (Vec<i8>, Vec<u64>)>,
-        Vec<i64>,
+        Box<dyn Fn(Vec<i32>, i8, u32, i8, u32) -> (Vec<i8>, Vec<u32>)>,
+        Vec<i32>,
     ),
 ) {
     (
@@ -95,13 +95,13 @@ pub fn approachgliss_on_interval(
     )
 }
 // pub fn first_takesall_on_interval(
-//     check_args: Vec<i64>,
-//     replace_args: Vec<i64>,
+//     check_args: Vec<i32>,
+//     replace_args: Vec<i32>,
 // ) -> (
-//     (Box<dyn Fn(Vec<i64>, i8, u64, i8, u64) -> bool>, Vec<i64>),
+//     (Box<dyn Fn(Vec<i32>, i8, u32, i8, u32) -> bool>, Vec<i32>),
 //     (
-//         Box<dyn Fn(Vec<i64>, i8, u64, i8, u64) -> (Vec<i8>, Vec<u64>)>,
-//         Vec<i64>,
+//         Box<dyn Fn(Vec<i32>, i8, u32, i8, u32) -> (Vec<i8>, Vec<u32>)>,
+//         Vec<i32>,
 //     ),
 // ) {
 //     (
